# Tekst-API/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Tekst-Dev",
    "description": "An online text research platform",
    "termsOfService": "None",
    "contact": {},
    "license": {
      "name": "AGPL-3.0-or-later",
      "url": "https://www.gnu.org/licenses/agpl-3.0.txt"
    },
    "version": "0.0.1"
  },
  "servers": [
    {
      "url": "http://127.0.0.1/api"
    }
  ],
  "paths": {
...
    "/contents": {
      "post": {
        "tags": [
          "contents"
        ],
        "summary": "Create content",
        "operationId": "createContent",
        "security": [
          {
            "APIKeyCookie": []
          },
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/DebugContentCreate"
                  },
                  {
                    "$ref": "#/components/schemas/PlaintextContentCreate"
                  }
                ],
                "discriminator": {
                  "propertyName": "resourceType",
                  "mapping": {
                    "debug": "#/components/schemas/DebugContentCreate",
                    "plaintext": "#/components/schemas/PlaintextContentCreate"
                  }
                },
                "title": "Content"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/DebugContentRead"
                    },
                    {
                      "$ref": "#/components/schemas/PlaintextContentRead"
                    }
                  ],
                  "discriminator": {
                    "propertyName": "resourceType",
                    "mapping": {
                      "debug": "#/components/schemas/DebugContentRead",
                      "plaintext": "#/components/schemas/PlaintextContentRead"
                    }
                  },
                  "title": "Response Create Content Contents Post"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "409": {
            "description": "Conflict"
          },
          "403": {
            "description": "Forbidden"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "contents"
        ],
        "summary": "Find contents",
        "description": "Returns a list of all resource contents matching the given criteria.\n\nRespects restricted resources and inactive texts.\nAs the resulting list may contain contents of different types, the\nreturned content objects cannot be typed to their precise resource content type.",
        "operationId": "findContents",
        "security": [
          {
            "APIKeyCookie": []
          },
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "res",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "example": "5eb7cf5a86d9755df3a6c593"
              },
              "description": "ID (or list of IDs) of resource(s) to return content data for",
              "default": [],
              "title": "Res"
            },
            "description": "ID (or list of IDs) of resource(s) to return content data for"
          },
          {
            "name": "location",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "example": "5eb7cf5a86d9755df3a6c593"
              },
              "description": "ID (or list of IDs) of location(s) to return content data for",
              "default": [],
              "title": "Location"
            },
            "description": "ID (or list of IDs) of location(s) to return content data for"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "description": "Return at most <limit> items",
              "default": 4096,
              "title": "Limit"
            },
            "description": "Return at most <limit> items"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "oneOf": [
                      {
                        "$ref": "#/components/schemas/DebugContentRead"
                      },
                      {
                        "$ref": "#/components/schemas/PlaintextContentRead"
                      }
                    ],
                    "discriminator": {
                      "propertyName": "resourceType",
                      "mapping": {
                        "debug": "#/components/schemas/DebugContentRead",
                        "plaintext": "#/components/schemas/PlaintextContentRead"
                      }
                    }
                  },
                  "title": "Response Find Contents Contents Get"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/contents/{id}": {
      "get": {
        "tags": [
          "contents"
        ],
        "summary": "Get content",
        "description": "A generic route for retrieving a content by ID from the database",
        "operationId": "getContent",
        "security": [
          {
            "APIKeyCookie": []
          },
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "5eb7cf5a86d9755df3a6c593",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/DebugContentRead"
                    },
                    {
                      "$ref": "#/components/schemas/PlaintextContentRead"
                    }
                  ],
                  "discriminator": {
                    "propertyName": "resourceType",
                    "mapping": {
                      "debug": "#/components/schemas/DebugContentRead",
                      "plaintext": "#/components/schemas/PlaintextContentRead"
                    }
                  },
                  "title": "Response Get Content Contents  Id  Get"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "contents"
        ],
        "summary": "Update content",
        "operationId": "updateContent",
        "security": [
          {
            "APIKeyCookie": []
          },
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "5eb7cf5a86d9755df3a6c593",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/DebugContentUpdate"
                  },
                  {
                    "$ref": "#/components/schemas/PlaintextContentUpdate"
                  }
                ],
                "discriminator": {
                  "propertyName": "resourceType",
                  "mapping": {
                    "debug": "#/components/schemas/DebugContentUpdate",
                    "plaintext": "#/components/schemas/PlaintextContentUpdate"
                  }
                },
                "title": "Updates"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/DebugContentRead"
                    },
                    {
                      "$ref": "#/components/schemas/PlaintextContentRead"
                    }
                  ],
                  "discriminator": {
                    "propertyName": "resourceType",
                    "mapping": {
                      "debug": "#/components/schemas/DebugContentRead",
                      "plaintext": "#/components/schemas/PlaintextContentRead"
                    }
                  },
                  "title": "Response Update Content Contents  Id  Patch"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "400": {
            "description": "Bad request"
          },
          "403": {
            "description": "Forbidden"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "contents"
        ],
        "summary": "Delete content",
        "operationId": "deleteContent",
        "security": [
          {
            "APIKeyCookie": []
          },
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "5eb7cf5a86d9755df3a6c593",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DebugContentCreate": {
        "properties": {
          "resourceId": {
            "type": "string",
            "title": "Resourceid",
            "description": "Resource ID",
            "example": "5eb7cf5a86d9755df3a6c593"
          },
          "resourceType": {
            "const": "debug",
            "title": "Resourcetype"
          },
          "locationId": {
            "type": "string",
            "title": "Locationid",
            "description": "Parent text location ID",
            "example": "5eb7cf5a86d9755df3a6c593"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment",
            "description": "Plaintext, potentially multiline comment on this content"
          },
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text",
            "description": "Text content of the debug content"
          }
        },
        "type": "object",
        "required": [
          "resourceId",
          "resourceType",
          "locationId"
        ],
        "title": "DebugContentCreate"
      },
      "DebugContentRead": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "example": "5eb7cf5a86d9755df3a6c593"
          },
          "resourceId": {
            "type": "string",
            "title": "Resourceid",
            "description": "Resource ID",
            "example": "5eb7cf5a86d9755df3a6c593"
          },
          "resourceType": {
            "const": "debug",
            "title": "Resourcetype"
          },
          "locationId": {
            "type": "string",
            "title": "Locationid",
            "description": "Parent text location ID",
            "example": "5eb7cf5a86d9755df3a6c593"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment",
            "description": "Plaintext, potentially multiline comment on this content"
          },
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text",
            "description": "Text content of the debug content"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "id",
          "resourceId",
          "resourceType",
          "locationId"
        ],
        "title": "DebugContentRead"
      },
      "DebugContentUpdate": {
        "properties": {
          "resourceId": {
            "anyOf": [
              {
                "type": "string",
                "example": "5eb7cf5a86d9755df3a6c593"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resourceid"
          },
          "resourceType": {
            "const": "debug",
            "title": "Resourcetype"
          },
          "locationId": {
            "anyOf": [
              {
                "type": "string",
                "example": "5eb7cf5a86d9755df3a6c593"
              },
              {
                "type": "null"
              }
            ],
            "title": "Locationid"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment",
            "description": "Plaintext, potentially multiline comment on this content"
          },
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text",
            "description": "Text content of the debug content"
          }
        },
        "type": "object",
        "required": [
          "resourceType"
        ],
        "title": "DebugContentUpdate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PlaintextContentCreate": {
        "properties": {
          "resourceId": {
            "type": "string",
            "title": "Resourceid",
            "description": "Resource ID",
            "example": "5eb7cf5a86d9755df3a6c593"
          },
          "resourceType": {
            "const": "plaintext",
            "title": "Resourcetype"
          },
          "locationId": {
            "type": "string",
            "title": "Locationid",
            "description": "Parent text location ID",
            "example": "5eb7cf5a86d9755df3a6c593"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment",
            "description": "Plaintext, potentially multiline comment on this content"
          },
          "text": {
            "type": "string",
            "maxLength": 102400,
            "minLength": 1,
            "title": "Text",
            "description": "Text content of the plaintext content"
          }
        },
        "type": "object",
        "required": [
          "resourceId",
          "resourceType",
          "locationId",
          "text"
        ],
        "title": "PlaintextContentCreate"
      },
      "PlaintextContentRead": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "example": "5eb7cf5a86d9755df3a6c593"
          },
          "resourceId": {
            "type": "string",
            "title": "Resourceid",
            "description": "Resource ID",
            "example": "5eb7cf5a86d9755df3a6c593"
          },
          "resourceType": {
            "const": "plaintext",
            "title": "Resourcetype"
          },
          "locationId": {
            "type": "string",
            "title": "Locationid",
            "description": "Parent text location ID",
            "example": "5eb7cf5a86d9755df3a6c593"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment",
            "description": "Plaintext, potentially multiline comment on this content"
          },
          "text": {
            "type": "string",
            "maxLength": 102400,
            "minLength": 1,
            "title": "Text",
            "description": "Text content of the plaintext content"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "id",
          "resourceId",
          "resourceType",
          "locationId",
          "text"
        ],
        "title": "PlaintextContentRead"
      },
      "PlaintextContentUpdate": {
        "properties": {
          "resourceId": {
            "anyOf": [
              {
                "type": "string",
                "example": "5eb7cf5a86d9755df3a6c593"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resourceid"
          },
          "resourceType": {
            "const": "plaintext",
            "title": "Resourcetype"
          },
          "locationId": {
            "anyOf": [
              {
                "type": "string",
                "example": "5eb7cf5a86d9755df3a6c593"
              },
              {
                "type": "null"
              }
            ],
            "title": "Locationid"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment",
            "description": "Plaintext, potentially multiline comment on this content"
          },
          "text": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 102400,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "resourceType"
        ],
        "title": "PlaintextContentUpdate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "APIKeyCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "tekstuserauth"
      },
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "auth/jwt/login"
          }
        }
      }
    }
  },
...
}